requirements.txt
streamlit
Pillow
matplotlib
numpy
trimesh
plotly
usd-core
